Fix integer loss of precision warning.

+++ evernote-sdk-ios/ENSDK/ENNoteRef.m
@@ -61,7 +61,7 @@
 
 - (void)encodeWithCoder:(NSCoder *)encoder
 {
-    [encoder encodeInt32:self.type forKey:@"type"];
+    [encoder encodeInt32:(int32_t)self.type forKey:@"type"];
     [encoder encodeObject:self.guid forKey:@"guid"];
     [encoder encodeObject:self.linkedNotebook forKey:@"linkedNotebook"];
 }
